requirements.txt
# ── Core ──────────────────────────────────────
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
stripe>=8.0.0

# ── AI ────────────────────────────────────────
anthropic>=0.40.0

# ── Data Science ──────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.5.0

# ── Data manipulation ─────────────────────────
pandas>=2.0.0

# ── Dev / Test ────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.4.0
